Move the pygmaps plotting code into a subdirectory to move to leaflet + OSM instead
Also fix the usages to match the new structure.
This is consistent with the planned directory structure.
All tests pass.
