+++ emission/analysis/classification/cleaning/gps_smoothing.py
@@ -7,8 +7,8 @@ import math
 
 # Our imports
 import emission.analysis.section_features as sf
-import emission.analysis.plotting.gmap_display as eapg
-import emission.analysis.plotting.pygmaps_modified as pygmaps
+import emission.analysis.plotting.gmaps.gmap_display as eapg
+import emission.analysis.plotting.gmaps.pygmaps_modified as pygmaps
 import emission.storage.decorations.useful_queries as taug
 import emission.core.get_database as edb
